+++ NEWS.md
@@ -1,3 +1,9 @@
+# version 0.4.1.9000
+
+* Upgrade Geocoder API version used in `geocode()` from v6.2 to v7 (closes [52](https://github.com/munterfinger/hereR/issues/52)).
+* Change default geometry in the return value of `geocode()` to position coordinates (display position) and return access coordinates (navigation position) as additional geometry list column. (closes [#53](https://github.com/munterfinger/hereR/issues/53)).
+
+
 # version 0.4.1
 
 * Change example and API mock data for `intermodal_route()` from Berlin to Switzerland, as the service is now also available there.